requirements.txt
alembic==1.13.1
apispec==6.6.0
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-Migrate==4.0.7
flask-smorest==0.44.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
marshmallow==3.21.1
packaging==24.0
psycopg2==2.9.9
python-dotenv==1.0.1
PyYAML==6.0.1
SQLAlchemy==2.0.28
typing_extensions==4.10.0
webargs==8.4.0
Werkzeug==3.0.1
